requirements.txt
gdown==4.7.1
matplotlib==3.5.3
numpy==1.21.6
opencv_python==4.8.0.74
opencv_python_headless==4.8.0.74
paddlehub==2.3.1
Pillow==9.5.0
Pillow==10.0.0
rembg==2.0.40
scikit_image==0.19.3
torch==1.13.1
torchvision==0.14.1
